+++ src/Application/Query/Happening/Webinar/Question/QuestionView.php
@@ -4,6 +4,9 @@ namespace Proximum\Vimeet\Application\Query\Happening\Webinar\Question;
 
 class QuestionView
 {
+    /** @var int */
+    public $questionId;
+
     /** @var string */
     public $questionContent;
 
@@ -25,15 +28,25 @@ class QuestionView
     /** @var string */
     public $createdAt;
 
+    /** @var int */
+    public $voteCount;
+
+    /** @var bool */
+    public $isLiked;
+
     public function __construct(
+        int $questionId,
         string $questionContent,
         string $firstName,
         string $lastName,
         ?string $position,
         ?string $avatar,
         ?string $sheetTitle,
-        string $createdAt
+        string $createdAt,
+        int $voteCount,
+        bool $isLiked
     ) {
+        $this->questionId = $questionId;
         $this->questionContent = $questionContent;
         $this->firstName = $firstName;
         $this->lastName = $lastName;
@@ -41,5 +54,7 @@ class QuestionView
         $this->avatar = $avatar;
         $this->sheetTitle = $sheetTitle;
         $this->createdAt = $createdAt;
+        $this->voteCount = $voteCount;
+        $this->isLiked = $isLiked;
     }
 }